third_party/WebKit/LayoutTests/imported/web-platform-tests/html/semantics/forms/the-input-element/month-expected.txt

CONSOLE WARNING: line 19: The specified value "invalid-month" does not conform to the required format.  The format is "yyyy-MM" where yyyy is year in four or more digits, and MM is 01-12.
CONSOLE WARNING: line 24: The specified value "13-06" does not conform to the required format.  The format is "yyyy-MM" where yyyy is year in four or more digits, and MM is 01-12.
CONSOLE WARNING: line 28: The specified value "2013-13" does not conform to the required format.  The format is "yyyy-MM" where yyyy is year in four or more digits, and MM is 01-12.
CONSOLE WARNING: line 29: The specified value "2013-00" does not conform to the required format.  The format is "yyyy-MM" where yyyy is year in four or more digits, and MM is 01-12.
This is a testharness.js-based test.
PASS month type support on input element 
FAIL The value attribute, if specified and not empty, must have a value that is a valid month string assert_equals: expected "2011-01" but got "1999-01"
FAIL The min attribute, if specified, must have a value that is a valid month string. assert_equals: expected "" but got "invalid_min"
FAIL The max attribute, if specified, must have a value that is a valid month string assert_equals: expected "2099-01" but got "2011-12"
PASS User agents must not allow the user to set the value to a non-empty string that is not a valid month string. 
PASS Month value can be empty string. 
PASS When value attribute has two digits year value, the value,which is invalid, must return empty string. 
PASS When value is set with invalid value, the value must return empty string. 
FAIL When value is given invalid value to non-empty valid string, the value must be same as before. assert_equals: expected "2013-05" but got ""
PASS When step attribute is given invalid value, it must ignore the invalid value and use defaul value instead. 
PASS Month should be <= 13: If the value of the element is not a valid month string, then set it to the empty string instead. 
PASS Month should be > 0: If the value of the element is not a valid month string, then set it to the empty string instead.> 
Harness: the test ran to completion.
